Split PadController into Create / Session / Lifecycle controllers
Closes #63.

PadController had grown to 347 LOC, 14 actions, and 8 injected
services — the file every new pad-related feature accreted onto.
Split along the three natural concerns from the architecture
review:

- PadCreateController (106 LOC) — create, createByParent,
  createFromTemplate, createFromUrl. One service dep
  (PadCreationService) on top of the shared base.
- PadSessionController (132 LOC) — open, openById, initialize,
  initializeById, metaById, resolveById. Three service deps
  (PadOpenService, PadInitializationService, PadMetadataService).
- PadLifecycleController (139 LOC) — trash, restore,
  recoverByFileId, syncById, syncStatusById, findOriginalByFileId.
  Three service deps (LifecycleService, PadSyncService,
  PadMetadataService).

Shared cross-cutting infrastructure (IUserSession, LoggerInterface,
IL10N, PadResponseService, PadControllerErrorMapper) plus the
helper methods (runForUser, requireUser, requireFileId,
requirePositiveInt, requireAccessMode, logError) move into a new
AbstractPadController base class. Each concrete controller's
constructor declares only the services it actually uses on top of
the base deps.

URL paths in appinfo/routes.php stay identical — only the handler
names change from 'pad#xxx' to 'padCreate#xxx' / 'padSession#xxx'
/ 'padLifecycle#xxx'. Public API unchanged.

PadControllerTest (890 LOC) split into three focused test files;
no behavior change. 395 tests stay green.
