+++ tests/phpunit/unit/PadLifecycleControllerTest.php
@@ -4,26 +4,20 @@ declare(strict_types=1);
 
 namespace OCA\EtherpadNextcloud\Tests\Unit;
 
-use OCA\EtherpadNextcloud\Controller\PadController;
 use OCA\EtherpadNextcloud\Controller\PadControllerErrorMapper;
+use OCA\EtherpadNextcloud\Controller\PadLifecycleController;
 use OCA\EtherpadNextcloud\Exception\PadAlreadyHasBindingException;
 use OCA\EtherpadNextcloud\Service\AppConfigService;
 use OCA\EtherpadNextcloud\Service\BindingService;
 use OCA\EtherpadNextcloud\Service\EtherpadClient;
 use OCA\EtherpadNextcloud\Service\ExternalPadExportFetcher;
 use OCA\EtherpadNextcloud\Service\LifecycleService;
-use OCA\EtherpadNextcloud\Service\PadCreationService;
 use OCA\EtherpadNextcloud\Service\PadFileLockRetryService;
 use OCA\EtherpadNextcloud\Service\PadFileService;
-use OCA\EtherpadNextcloud\Service\PadInitializationService;
 use OCA\EtherpadNextcloud\Service\PadMetadataService;
-use OCA\EtherpadNextcloud\Service\PadOpenService;
 use OCA\EtherpadNextcloud\Service\PadResponseService;
-use OCA\EtherpadNextcloud\Service\PadSessionService;
 use OCA\EtherpadNextcloud\Service\PadSyncService;
 use OCA\EtherpadNextcloud\Service\ParsedPadFile;
-use OCA\EtherpadNextcloud\Service\SnapshotExtractor;
-use OCA\EtherpadNextcloud\Service\SnapshotHtmlSanitizer;
 use OCA\EtherpadNextcloud\Service\UserNodeResolver;
 use OCA\EtherpadNextcloud\Util\PathNormalizer;
 use OCP\AppFramework\Http;
@@ -37,353 +31,7 @@ use OCP\Lock\LockedException;
 use PHPUnit\Framework\TestCase;
 use Psr\Log\LoggerInterface;
 
-class PadControllerTest extends TestCase {
-	public function testCreateReturnsUnauthorizedWhenNoUserSession(): void {
-		$userSession = $this->createMock(IUserSession::class);
-		$userSession->method('getUser')->willReturn(null);
-
-		$controller = $this->buildController($this->createMock(IRequest::class), $userSession);
-		$response = $controller->create('/Test.pad');
-
-		$this->assertSame(Http::STATUS_UNAUTHORIZED, $response->getStatus());
-		$this->assertSame('Authentication required.', $response->getData()['message']);
-	}
-
-	public function testCreateByParentReturnsUnauthorizedWhenNoUserSession(): void {
-		$userSession = $this->createMock(IUserSession::class);
-		$userSession->method('getUser')->willReturn(null);
-
-		$controller = $this->buildController($this->createMock(IRequest::class), $userSession);
-		$response = $controller->createByParent(123, 'Test');
-
-		$this->assertSame(Http::STATUS_UNAUTHORIZED, $response->getStatus());
-		$this->assertSame('Authentication required.', $response->getData()['message']);
-	}
-
-	public function testCreateByParentRejectsInvalidParentFolderId(): void {
-		$user = $this->createMock(IUser::class);
-		$userSession = $this->createMock(IUserSession::class);
-		$userSession->method('getUser')->willReturn($user);
-
-		$controller = $this->buildController($this->createMock(IRequest::class), $userSession);
-		$response = $controller->createByParent(0, 'Test');
-
-		$this->assertSame(Http::STATUS_BAD_REQUEST, $response->getStatus());
-		$this->assertSame('Invalid parentFolderId.', $response->getData()['message']);
-	}
-
-	public function testCreateByParentRejectsInvalidAccessMode(): void {
-		$user = $this->createMock(IUser::class);
-		$userSession = $this->createMock(IUserSession::class);
-		$userSession->method('getUser')->willReturn($user);
-
-		$controller = $this->buildController($this->createMock(IRequest::class), $userSession);
-		$response = $controller->createByParent(12, 'Test', 'invalid');
-
-		$this->assertSame(Http::STATUS_BAD_REQUEST, $response->getStatus());
-		$this->assertSame('Invalid accessMode. Use public or protected.', $response->getData()['message']);
-	}
-
-	public function testOpenByIdRejectsInvalidFileId(): void {
-		$user = $this->createMock(IUser::class);
-		$userSession = $this->createMock(IUserSession::class);
-		$userSession->method('getUser')->willReturn($user);
-
-		$controller = $this->buildController($this->createMock(IRequest::class), $userSession);
-		$response = $controller->openById(0);
-
-		$this->assertSame(Http::STATUS_BAD_REQUEST, $response->getStatus());
-		$this->assertSame('Invalid file ID.', $response->getData()['message']);
-	}
-
-	public function testOpenByIdRetriesLockedReadAndEventuallySucceeds(): void {
-		$user = $this->createConfiguredMock(IUser::class, [
-			'getUID' => 'alice',
-			'getDisplayName' => 'Alice',
-		]);
-		$userSession = $this->createConfiguredMock(IUserSession::class, ['getUser' => $user]);
-		$file = $this->buildPadFileNode();
-		$file->expects($this->exactly(3))
-			->method('getContent')
-			->willReturnCallback(static function (): string {
-				static $call = 0;
-				$call++;
-				if ($call < 3) {
-					throw new LockedException('locked');
-				}
-				return 'frontmatter';
-			});
-
-		$rootFolder = $this->createMock(IRootFolder::class);
-		$rootFolder->method('getById')->with(138)->willReturn([$file]);
-
-		$padFileService = $this->createMock(PadFileService::class);
-		$padFileService->expects($this->once())
-			->method('readPad')
-			->with('frontmatter')
-			->willReturn(new ParsedPadFile(
-				frontmatter: [
-					'pad_id' => 'g.ABCDEFGHIJKLMNOP$pad-1',
-					'access_mode' => BindingService::ACCESS_PUBLIC,
-				],
-				body: '',
-				padId: 'g.ABCDEFGHIJKLMNOP$pad-1',
-				accessMode: BindingService::ACCESS_PUBLIC,
-				padUrl: '',
-				isExternal: false,
-			));
-
-		$bindingService = $this->createMock(BindingService::class);
-		$bindingService->expects($this->once())
-			->method('assertConsistentMapping')
-			->with(138, 'g.ABCDEFGHIJKLMNOP$pad-1', BindingService::ACCESS_PUBLIC);
-
-		$etherpadClient = $this->createMock(EtherpadClient::class);
-		$etherpadClient->expects($this->once())
-			->method('buildPadUrl')
-			->with('g.ABCDEFGHIJKLMNOP$pad-1')
-			->willReturn('https://pad.example.test/p/g.ABCDEFGHIJKLMNOP$pad-1');
-
-		$appConfigService = $this->createMock(AppConfigService::class);
-		$appConfigService->expects($this->once())
-			->method('getSyncIntervalSeconds')
-			->willReturn(30);
-
-		$urlGenerator = $this->createMock(IURLGenerator::class);
-		$urlGenerator->method('linkToRoute')
-			->willReturnMap([
-				['etherpad_nextcloud.pad.syncById', ['fileId' => 138], '/sync/138'],
-				['etherpad_nextcloud.pad.syncStatusById', ['fileId' => 138], '/sync-status/138'],
-		]);
-		$logger = $this->createMock(LoggerInterface::class);
-		$padPaths = new PathNormalizer();
-		$userNodeResolver = new UserNodeResolver($rootFolder);
-		$lockRetryService = $this->buildNoSleepLockRetryService();
-		$padOpenService = new PadOpenService(
-			$padFileService,
-			$padPaths,
-			$userNodeResolver,
-			$lockRetryService,
-			$bindingService,
-			$etherpadClient,
-			$this->createMock(ExternalPadExportFetcher::class),
-			$this->createMock(PadSessionService::class),
-			new SnapshotExtractor($padFileService, new SnapshotHtmlSanitizer()),
-			$logger,
-		);
-		$l10n = $this->createMock(\OCP\IL10N::class);
-		$l10n->method('t')->willReturnCallback(static fn (string $text, array $params = []): string => $text);
-		$padResponseService = new PadResponseService($urlGenerator, $appConfigService, $l10n);
-
-		$controller = new PadController(
-			'etherpad_nextcloud',
-			$this->createMock(IRequest::class),
-			$userSession,
-			$logger,
-			$l10n,
-			$this->createMock(PadCreationService::class),
-			$this->createMock(PadInitializationService::class),
-			$this->createMock(PadMetadataService::class),
-			$padOpenService,
-			$this->createMock(PadSyncService::class),
-			$this->createMock(LifecycleService::class),
-			$padResponseService,
-			new PadControllerErrorMapper($padResponseService, $logger),
-		);
-
-		$response = $controller->openById(138);
-
-		$this->assertSame(Http::STATUS_OK, $response->getStatus());
-		$this->assertSame('https://pad.example.test/p/g.ABCDEFGHIJKLMNOP$pad-1', $response->getData()['url']);
-	}
-
-	public function testOpenByIdReturnsRetryableErrorWhenReadRemainsLocked(): void {
-		$user = $this->createConfiguredMock(IUser::class, [
-			'getUID' => 'alice',
-			'getDisplayName' => 'Alice',
-		]);
-		$userSession = $this->createConfiguredMock(IUserSession::class, ['getUser' => $user]);
-		$file = $this->buildPadFileNode();
-		$file->expects($this->exactly(4))
-			->method('getContent')
-			->willThrowException(new LockedException('still locked'));
-
-		$rootFolder = $this->createMock(IRootFolder::class);
-		$rootFolder->method('getById')->with(138)->willReturn([$file]);
-
-		$controller = $this->buildController(
-			$this->createMock(IRequest::class),
-			$userSession,
-			rootFolder: $rootFolder,
-		);
-		$response = $controller->openById(138);
-
-		$this->assertSame(Http::STATUS_SERVICE_UNAVAILABLE, $response->getStatus());
-		$this->assertSame('Pad file is temporarily locked. Please retry.', $response->getData()['message']);
-		$this->assertTrue($response->getData()['retryable']);
-	}
-
-	public function testOpenByIdReturnsExternalPadUrlForExternalPads(): void {
-		$user = $this->createConfiguredMock(IUser::class, [
-			'getUID' => 'alice',
-			'getDisplayName' => 'Alice',
-		]);
-		$userSession = $this->createConfiguredMock(IUserSession::class, ['getUser' => $user]);
-		$file = $this->buildPadFileNode();
-
-		$rootFolder = $this->createMock(IRootFolder::class);
-		$rootFolder->method('getById')->with(138)->willReturn([$file]);
-
-		$frontmatter = [
-			'pad_id' => 'ext.abc123',
-			'access_mode' => BindingService::ACCESS_PUBLIC,
-			'pad_url' => 'https://pad.portal.fzs.de/p/Test',
-			'pad_origin' => 'https://pad.portal.fzs.de',
-			'remote_pad_id' => 'Test',
-		];
-
-		$padFileService = $this->createMock(PadFileService::class);
-		$padFileService->expects($this->once())
-			->method('readPad')
-			->with('frontmatter')
-			->willReturn(new ParsedPadFile(
-				frontmatter: $frontmatter,
-				body: '',
-				padId: 'ext.abc123',
-				accessMode: BindingService::ACCESS_PUBLIC,
-				padUrl: 'https://pad.portal.fzs.de/p/Test',
-				isExternal: true,
-			));
-		$padFileService->expects($this->once())
-			->method('getTextSnapshotForRestore')
-			->with('frontmatter')
-			->willReturn("External snapshot\nSecond line");
-		$padFileService->expects($this->once())
-			->method('getHtmlSnapshotForRestore')
-			->with('frontmatter')
-			->willReturn('<h1>External</h1><script>bad()</script>');
-
-		$bindingService = $this->createMock(BindingService::class);
-		$bindingService->expects($this->never())->method('assertConsistentMapping');
-
-		$etherpadClient = $this->createMock(EtherpadClient::class);
-		$etherpadClient->expects($this->never())->method('buildPadUrl');
-		$externalPadExportFetcher = $this->createMock(ExternalPadExportFetcher::class);
-		$externalPadExportFetcher->expects($this->once())
-			->method('normalizeAndValidateExternalPublicPadUrl')
-			->with('https://pad.portal.fzs.de/p/Test')
-			->willReturn(['pad_url' => 'https://pad.portal.fzs.de/p/Test']);
-
-		$appConfigService = $this->createMock(AppConfigService::class);
-		$appConfigService->expects($this->once())
-			->method('getSyncIntervalSeconds')
-			->willReturn(30);
-
-		$urlGenerator = $this->createMock(IURLGenerator::class);
-		$urlGenerator->method('linkToRoute')
-			->willReturnMap([
-				['etherpad_nextcloud.pad.syncById', ['fileId' => 138], '/sync/138'],
-				['etherpad_nextcloud.pad.syncStatusById', ['fileId' => 138], '/sync-status/138'],
-		]);
-		$logger = $this->createMock(LoggerInterface::class);
-		$padPaths = new PathNormalizer();
-		$userNodeResolver = new UserNodeResolver($rootFolder);
-		$lockRetryService = $this->buildNoSleepLockRetryService();
-		$padOpenService = new PadOpenService(
-			$padFileService,
-			$padPaths,
-			$userNodeResolver,
-			$lockRetryService,
-			$bindingService,
-			$etherpadClient,
-			$externalPadExportFetcher,
-			$this->createMock(PadSessionService::class),
-			new SnapshotExtractor($padFileService, new SnapshotHtmlSanitizer()),
-			$logger,
-		);
-		$l10n = $this->createMock(\OCP\IL10N::class);
-		$l10n->method('t')->willReturnCallback(static fn (string $text, array $params = []): string => $text);
-		$padResponseService = new PadResponseService($urlGenerator, $appConfigService, $l10n);
-
-		$controller = new PadController(
-			'etherpad_nextcloud',
-			$this->createMock(IRequest::class),
-			$userSession,
-			$logger,
-			$l10n,
-			$this->createMock(PadCreationService::class),
-			$this->createMock(PadInitializationService::class),
-			$this->createMock(PadMetadataService::class),
-			$padOpenService,
-			$this->createMock(PadSyncService::class),
-			$this->createMock(LifecycleService::class),
-			$padResponseService,
-			new PadControllerErrorMapper($padResponseService, $logger),
-		);
-
-		$response = $controller->openById(138);
-
-		$this->assertSame(Http::STATUS_OK, $response->getStatus());
-		$this->assertSame('https://pad.portal.fzs.de/p/Test', $response->getData()['url']);
-		$this->assertTrue($response->getData()['is_external']);
-		$this->assertSame('https://pad.portal.fzs.de/p/Test', $response->getData()['pad_url']);
-		$this->assertSame('https://pad.portal.fzs.de/p/Test', $response->getData()['original_pad_url']);
-		$this->assertSame("External snapshot\nSecond line", $response->getData()['snapshot_text']);
-		$this->assertSame('<h1>External</h1>', $response->getData()['snapshot_html']);
-	}
-
-	public function testMetaByIdRejectsInvalidFileId(): void {
-		$user = $this->createMock(IUser::class);
-		$userSession = $this->createMock(IUserSession::class);
-		$userSession->method('getUser')->willReturn($user);
-
-		$controller = $this->buildController($this->createMock(IRequest::class), $userSession);
-		$response = $controller->metaById(0);
-
-		$this->assertSame(Http::STATUS_BAD_REQUEST, $response->getStatus());
-		$this->assertSame('Invalid file ID.', $response->getData()['message']);
-	}
-
-	public function testMetaByIdReturnsRetryableErrorWhenReadRemainsLocked(): void {
-		$user = $this->createConfiguredMock(IUser::class, ['getUID' => 'alice']);
-		$userSession = $this->createConfiguredMock(IUserSession::class, ['getUser' => $user]);
-		$file = $this->buildPadFileNode();
-		$file->expects($this->exactly(4))
-			->method('getContent')
-			->willThrowException(new LockedException('still locked'));
-
-		$rootFolder = $this->createMock(IRootFolder::class);
-		$rootFolder->method('getById')->with(138)->willReturn([$file]);
-
-		$controller = $this->buildController(
-			$this->createMock(IRequest::class),
-			$userSession,
-			rootFolder: $rootFolder,
-		);
-		$response = $controller->metaById(138);
-
-		$this->assertSame(Http::STATUS_SERVICE_UNAVAILABLE, $response->getStatus());
-		$this->assertSame('Pad file is temporarily locked. Please retry.', $response->getData()['message']);
-		$this->assertTrue($response->getData()['retryable']);
-	}
-
-	public function testResolveByIdReturnsGenericServerErrorForUnexpectedFailures(): void {
-		$user = $this->createConfiguredMock(IUser::class, ['getUID' => 'alice']);
-		$userSession = $this->createConfiguredMock(IUserSession::class, ['getUser' => $user]);
-		$rootFolder = $this->createMock(IRootFolder::class);
-		$rootFolder->method('getById')->with(138)->willThrowException(new \RuntimeException('Storage offline.'));
-
-		$controller = $this->buildController(
-			$this->createMock(IRequest::class),
-			$userSession,
-			rootFolder: $rootFolder,
-		);
-		$response = $controller->resolveById(138);
-
-		$this->assertSame(Http::STATUS_INTERNAL_SERVER_ERROR, $response->getStatus());
-		$this->assertSame('Could not resolve pad file.', $response->getData()['message']);
-	}
-
+class PadLifecycleControllerTest extends TestCase {
 	public function testSyncByIdRejectsInvalidFileId(): void {
 		$user = $this->createMock(IUser::class);
 		$userSession = $this->createMock(IUserSession::class);
@@ -815,7 +463,7 @@ class PadControllerTest extends TestCase {
 		?EtherpadClient $etherpadClient = null,
 		?LifecycleService $padLifecycleOperations = null,
 		?ExternalPadExportFetcher $externalPadExportFetcher = null,
-	): PadController {
+	): PadLifecycleController {
 		$resolvedRootFolder = $rootFolder ?? $this->createMock(IRootFolder::class);
 		$resolvedEtherpadClient = $etherpadClient ?? $this->createMock(EtherpadClient::class);
 		$resolvedExternalPadExportFetcher = $externalPadExportFetcher ?? $this->createMock(ExternalPadExportFetcher::class);
@@ -826,18 +474,6 @@ class PadControllerTest extends TestCase {
 		$userNodeResolver = new UserNodeResolver($resolvedRootFolder);
 		$lockRetryService = $this->buildNoSleepLockRetryService();
 		$padMetadataService = new PadMetadataService($resolvedPadFileService, $padPaths, $userNodeResolver, $lockRetryService, $resolvedEtherpadClient, $resolvedExternalPadExportFetcher, $resolvedBindingService, $logger);
-		$padOpenService = new PadOpenService(
-			$resolvedPadFileService,
-			$padPaths,
-			$userNodeResolver,
-			$lockRetryService,
-			$resolvedBindingService,
-			$resolvedEtherpadClient,
-			$resolvedExternalPadExportFetcher,
-			$this->createMock(PadSessionService::class),
-			new SnapshotExtractor($resolvedPadFileService, new SnapshotHtmlSanitizer()),
-			$logger,
-		);
 		$padSyncService = new PadSyncService($resolvedPadFileService, $userNodeResolver, $lockRetryService, $resolvedBindingService, $resolvedEtherpadClient, $resolvedExternalPadExportFetcher, $logger);
 		$padLifecycleOperations = $padLifecycleOperations
 			?? $this->createMock(LifecycleService::class);
@@ -857,20 +493,17 @@ class PadControllerTest extends TestCase {
 		$l10n = $this->createMock(\OCP\IL10N::class);
 		$l10n->method('t')->willReturnCallback(static fn (string $text, array $params = []): string => $text);
 		$padResponseService = new PadResponseService($urlGenerator, $appConfigService, $l10n);
-		return new PadController(
+		return new PadLifecycleController(
 			'etherpad_nextcloud',
 			$request,
 			$userSession,
 			$logger,
 			$l10n,
-			$this->createMock(PadCreationService::class),
-			$this->createMock(PadInitializationService::class),
-			$padMetadataService,
-			$padOpenService,
-			$padSyncService,
-			$padLifecycleOperations,
 			$padResponseService,
 			new PadControllerErrorMapper($padResponseService, $logger),
+			$padLifecycleOperations,
+			$padSyncService,
+			$padMetadataService,
 		);
 	}
